updated and unified ui for comparison and wishlist
