+++ project-base/storefront/components/Pages/ProductComparison/ProductComparisonHead.tsx
@@ -1,4 +1,3 @@
-import { ProductComparisonButtonRemoveAll } from './ProductComparisonButtonRemoveAll';
 import { ProductComparisonHeadItem } from './ProductComparisonHeadItem';
 import { TypeProductInProductListFragment } from 'graphql/requests/productLists/fragments/ProductInProductListFragment.generated';
 import { useComparison } from 'utils/productLists/comparison/useComparison';
@@ -13,9 +12,8 @@ export const ProductComparisonHead: FC<ProductComparisonHeadProps> = ({ compared
     return (
         <thead>
             <tr id="js-table-compare-head">
-                <td className="z-above bg-table-bg-default sticky left-0 max-w-[205px] min-w-[115px] pr-3 align-top sm:w-52 sm:min-w-[211px] md:min-w-[256px] lg:w-72">
-                    <ProductComparisonButtonRemoveAll />
-                </td>
+                <td className="z-above bg-table-bg-default sticky left-0 max-w-[205px] min-w-[115px] pr-3 align-top sm:w-52 sm:min-w-[211px] md:min-w-[256px] lg:w-72" />
+
                 {comparedProducts.map((product, index) => (
                     <ProductComparisonHeadItem
                         key={`head-${product.uuid}`}
