+++ project-base/storefront/public/locales/cs/common.json
@@ -482,7 +482,7 @@
     "Related Products": "Související produkty",
     "Relevance": "Relevance",
     "Remove": "Odstranit",
-    "Remove all": "Odebrat vše",
+    "Remove all from comparison": "Odebrat vše z porovnání",
     "Remove all from wishlist": "Odebrat vše z oblíbených produktů",
     "Remove file": "Odebrat soubor",
     "Remove from cart": "Odstranit z košíku",
