+++ project-base/storefront/public/locales/en/common.json
@@ -452,7 +452,7 @@
     "Related Products": "Related Products",
     "Relevance": "Relevance",
     "Remove": "Remove",
-    "Remove all": "Remove all",
+    "Remove all from comparison": "Remove all from comparison",
     "Remove all from wishlist": "Remove all from wishlist",
     "Remove file": "Remove file",
     "Remove from cart": "Remove from cart",
