+++ project-base/storefront/public/locales/sk/common.json
@@ -482,7 +482,7 @@
     "Related Products": "Súvisiace produkty",
     "Relevance": "Relevantnosť",
     "Remove": "Odstrániť",
-    "Remove all": "Odstrániť všetko",
+    "Remove all from comparison": "Odstrániť všetko z porovnania",
     "Remove all from wishlist": "Odstrániť všetko z obľúbených produktov",
     "Remove file": "Odstrániť súbor",
     "Remove from cart": "Odstrániť z košíka",
